Updated Separated Code
Current updated code as of 3/23. Everything is working together except for the LCD screen. I have the timeSignature() function controlling everything, and when it is the only function called in loop(), it works great. But when LCD_Menu() is called with it, the LED's blink slower than they should.  Read the comments in the C_Tempo_Function for more details.